+++ src/Utils/CorePHPUnit.php
@@ -3,12 +3,16 @@ declare(strict_types=1);
 
 namespace lucatume\WPBrowser\Utils;
 
-class CorePHPunit
+class CorePHPUnit
 {
     public static function path(?string $string = null): string
     {
         $path = dirname(__DIR__, 2) . '/includes/core-phpunit';
 
         return $string ? $path . '/' . ltrim($string, '\\/') : $path;
+    }
+
+    public static function includes(string $path)
+    {
     }
 }