+++ src/Utils/Dates.php
@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 /**
  * Provides methods to interact and build dates.
  *